Add files via upload
These additions make the game relatively self sufficient. Malicious input is not protected for yet. This should allow a relatively functional playing experience. Catches for malicious users as well as help functions must be added in the future. Considered bells and whistles could be items, heroes, or monsters. The best to start with however would be a function to crate events in rooms based on a users choice of direction.